Fixed an issue with the file timestamp not being setup for a TFile when getting the file info on the Mac.  This fixes a problem where files were continually being re-output when they didn't need to be.
NOTE: I've copied the same code to the iPod version but disabled it as all files on the ipod will be final .asset files therefore files cannot be out of date as such.  If we ever add the ability to output the source files via the iPod then it is there ready to be used.