Fixed an issue with the drag and scroll intermittently working.  Looks like this was down to the use of the widget type.  The drag and scroll were sending through TLGui_WidgetActionType_Down instead of TLGui_WidgetActionType_Move when dragging objects/scroll menus which seemed to sometimes work but sometimes not.
Additionally in the widget ProcessClick() it was errornously checking the click GetActionRef() for TLGui_WidgetActionType_Move instead of the GetActionType() so the click would never register as a move.  

TBH I have no idea how the drags actually worked at all seeing as this code has been like this for quite some time!? :S  Anyway, seems ok now after testing on the Mac and the iPod with Rocket.

Moved the hack from yesterday for the cursor pos to a define at the top of the file so that it's easier to edit and also so I can find it when I eventually look into a better fix, if there is one.