Added a LifeChangeType ref to the TSceneNode_Object.  This can be used to determine the type of damage done i.e. electircal, fire, etc etc.  The type can be sent as part of the ChangeLife or SetLife messages.
NOTE: Depending on requirements, this may change in the future to to be a full on message that is passed on to the OnLifeChange() and OnDeath() routines rather than being stored on the node so that more verbose information can be accessed, such as who caused the life change, position of intersection of collision if any or any other pertinent details.  For now the type should suffice.