Added the TWidgetFactory.cpp and .h files to the IPod and Mac TootleGame library.  NOTE: Needs adding to the PC solution.
Implemented the widget factory for the widget manager.  This is the first phase for moving some of the input processing from the widgets onto the widget manager.

Added a new constructor to all widget classes for use with the widget factory that takse an instance ref and type ref.  The widget factory is setup to store the TWidget once instanced so it can be looked up via the TWidgetManager.

Added an Initialise(TMessage&) routine to the TWidget base class. This will be the main initialisation routine which will need to be sent after widget has been created with the appropriate data.

Added SetProperty(TMessage&) to the widget classes as appropriate.  This will eventually be the main routine for setting widget properties much like how the nodes work.  This is called prioir to the Initialise() when an initialise message is received.

Added TKeyArray< TRef, TArray<TRef> > to the widget manager that will be used for keeping track of the widgets using a TRenderTarget ref for the key.  It will eventually be sorted b render target z and the array of widget refs will also be ordered by z.