Added a FindChildOfType(TRefRef TypeRef) to the TGraphNode class, to find a child node with a specific type ref.  Useful for when instancing child nodes via messaing/XML without keeping TRef's of the child node(s).
NOTE: This will return the first one found (if any) that match the type specified, so this may need expanding in the future to return an array instead, depending on usage.

Added a RemoveWidget(TRefRef) routine to the widget manager to allow removal of a widget created via the manager.