Minor change to the TMessageQueue, TPublisher and TSubscriber changing their destructor and shutdown() routines to be protected.
This ensures we can't delete objects using pointers to these types which could have undefined behaviour depending on usage of the class in the class hierarchy of an object.

The shutdown can now only be called via the class itself or any derived classes that may need to explcitly call this routine.