Added a templated GetApplication() routine to the TGame class.
Updated the menu virtual routine OnMenuItemExecuted() to have a TBinaryTree& passed to it.  This is so we can process any extra data from the XML after the command has been executed.  

NOTE: I have mainly added this to cover *all* menu commands because currently the open and close commands are handled specially compared to other commands and they do not pass the extra data through like with the ExecuteCommand because they are done via a delayed QueuedCommand reference. Adding the passing of this data would involve having a QueuedCommandData tree I think but for my needs processing the extra data after the command has executed will work ok.

Added the overloading of the TMessageQueue to the TSchemeEditor.  This allows us to queue up messages on the editor using the public QueueMessage(TMessage&) call.