Updated the iPod and Mac code in line with changes to the TString's now using unicode char's.
The wchar_t type in GCC is 4-bytes rather than 2-bytes so to go between TStrings and the native NSString's I've added platform specific MacString and IPodString files along with a NSString* TLString::ConvertToUnicharString(const TString& ) routine.  

This is only temporary as the TChar16 will need to be 2-bytes to match the PC and avoid any binary incompatibilites.  See TRAC ticket #114 for info.