Consolidated the Platform::PrintToBuffer(TString& ) routines across all platforms to fix issues with the strings not being output to the console window.
They are now all the same and do not use the Appendf() routine as it's not necessary - the PC version already did this so the changes have been merged from that version.  

Also updated the use of the newline terminator to ensure the buffer size is correct - it was assuming the terminator character was a byte in length before which was wrong.  NOTE: This could be further improved with a static newline char perhaps rather than creating the TTempString?