Added a DEBUG_CONSOLE_BUFFER_SIZE for each platform so that the buffered console output can be different per platform if needs be at some stage.
Updated the PrintToBuffer(TString&) routine to remove the creaiton fo the TTempString newline("\n") and also use the g_ConsoleBuffer's alloc size to determine if the buffer needs flushing.

Corrected a bug in the TFixedArray that would not use it's virtual GetAllocSize() due to it not being declared as const like the one in TArray.