Updated the Scheme editor and menu wrapper to use the new widget creation.  These will now simply keep track of the widgets via TRef's rather than widget pointers.
FYI there is on one more place in the engine that needs updating to use the new system for widget creation and that is the editor node widget attachment and removal.  This needs some extra support to test the widget render node ref so I've left this for now.

Set the old TWidget constructors as deprecated to make it easier to spot and correct old usages.  The Shutdown is also now deprecated as it is only called from the destructor and I will remove it once the issues with the TWidget arrays are sorted out completely.

Updated the widget system to use TWidget pointers in the key array rather than TRefs and the class factory now doesn't store the TWidget pointers. This is needed because the key array allows us to have duplicate TRef's for the widget across different groups (the group is the render target key value) but the class factory is one big list so accessing a widget with a specific ref could alter the wrong one if there are more than one with the same ref.  Storing the TWidget TPtr's in the key array resolves this potential problem.

Updated the widget manager public routines to require the group ref as part of any widget access to ensure the correct widget is used.

Added a widget cache to the widget manager to improve performance when searching for widgets within the widget system.  This is disabled for now whilst the issues with the keyarray tptrarraay is looked into further.