Added a network CreateConnection() wrapper that will create a platform specific network connection.  This simplifies the gameside code to create a connection. Eventually this should be part of some network or download manager to ensure there is only ever one.
Fixed various crashes when using the iPod http connection. One issue was having the delegate as part of the connection class where the delegate was being released prematurely and another issue was with starting the connection immediately and triggering the start routines - this caused multiple threads to be created internally where one would eventually timeout causing a crash and also meant it was difficult to see why it was crashing as the stack would be the second thread not the game thread.  

I'll propogate the changes from the iPod into the Mac code in due course as they work in almost the same way.