Found an old engine of mine. It's pretty readable so it's probably about 5/6 years old :)
I think most of it has been superceeded now, but it has a skeleton/bone system, software & hardware shader (NV shader, but I'm sure I wrote an arb one...) which might be of some use to use.

Committing so it's easy to access for me. I may have a look through it soon to see if there's anything to extract. Once I've been through it I'll delete it again :)