Added initial work on accessing the camera on the iPhone.  This will only compile with iPhone OS 3.1 SDK or above and as yet is untested.
Added support for the camera image to be used as a texture on the Mac.  
The camera delegate will broadcast the image as a message which is intercepted by the camera manager.  The camera manager can create and reference a texture asset for use as the camera image texture, which gets updated when the image is sent to the camera manager.  Any polygons referencing the camera image texture will automatically get updated with the camera view.  

This has meant the following additional changes to other parts of the engine:

* Texture asset now has a flag to say whether it is changed at runtime
* OpenGL render will test texture changed flag and re-upload the texture
* Added a specific routine to the texture asset class to update the texture data - this was becuase using the ImportData() (that I wanted to use) would never set the data correctly so added routine for now until I can investigate and get it working via the ImportData.

Eventually the texture flag should be removed and replaced with a message to get the render system to keep track of textures that need updating within the renderer, rather than teting/editing flags ont eh texture asset.

The following are current issues that need further investigation:

* Image will flicker white.  I think this is down to deleting the texture and re-uploading it once it gets to the render stage but I can't be certain.
* Sometimes crashes in the memory system when exiting and comes from the release of the captured image.  Presumably this is because the image is buffered so needs to be cleared via some obj-c code somewhere at the Mac level.
* If the camera view window isn't visible (hidden or offscreen) the image from the camera isn's sent via the delegate.

Added some maths routines to handle testing for powers of two and getting the next power of two.  These have replaced some specific checks in the texture asset class.