Changed the FORCEINLINE of the TestVibration routine and moved the body of code into the .mm file.
Not sure about this one.  It used to build ok but now doesn't unless the body of the routine (which is empty) is in the .cpp/.mm file?  It produces duplicate symbol compilation errors.

Must be some build property change as it only started after we changed the xcconfig references.  I've noticed there are a few others that have the same issue when they should inline perfectly well.