After searching for a problem on the mac with the scream audio not working correctly in the testapp this turned out to be the distance processing, so now by default all games have the 3D processing disabled (effectively making them 2D) whilst I investigate further.
Added low level routines to setup the distance model and doppler effect for all platforms.

Updated the initialisation (enable) of the low level system to NOT automatically set the distance model, doppler effect or listener anymore - this is now done by the audio graph when enabling the system from a higher level and uses data on the audio graph.  This will also mean the setup is re-initialised to the last (correct) data set between being disabled and enabled, whereas previously it would revert to whatever it started with.

The distance model and doppler can now be set as properties via the messaging and therefore we can change these by setting the appropriate values per project using the  .options file.  I'll add an example to one of the projects in the near future.  By default all games will start out 2D which will be useful to ensure the audio on any bootup screens works at the start.

Added a new TAudioDistance.h file to wrap any distance model definitions/routines that will be cross-platform.