Added platform specific maths files so we can now specialise maths routines per platform.
Moved various platform specific maths routines into the new platform specific files.

Renamed the sinf and cosf routines that could potentially use a lookup table to be a specific 'quick' version - Qsinf and Qcosf.  Added the sinf and cosf to be the standard system versions by default.
Additionally I've ported the code for setting up the lookup tables into the Ipod version of the platform specific code only as I suspect this would only be needed for the Ipod?  If not it can always be transfered to the others as well.

NOTE: If the default behaviour should be the quick version we could change this so that there is a system one for when we want to access the system versions instead where sinf and cosf would utilise the lookup table(s). 

Seperated off the TAxisAngle, TEuler, TQuaternion and TMatrix classes into their own files for easier maintenance and to simplify the generic TLMaths.cpp and .h files.  It should also reduce the build dependency a bit.