Moved various build options into more appropriate places and removed some unnecessary duplicates.
Added new IPodDebug and IPodRelease xcconfig files.  This is more inline with Mac version and should be used for all Debug/Release options across the platform regardless of project build type.

Re-enabled dead code stripping for the IPod build by default.

Updated the debug symbol generation to be full.  This is a prerequisite for dead code stripping to work correctly.  In release these should get stripped out when linking.

Updated the symbols private extern to be YES for the IPod as we are currently doing a static link with the C++ runtime which requires this setting.  According to the documentation it is actually better practice to use this option and make symbols explicitly visible. NOTE: This may or may not produce warnings for the wxwidgets build(s).
To facilitate any required symbol visibility in the future I've added an EXPORT and IMPORT macro to TLTypes.h, that would be needed if we were to build dynamic libraries instead of static ones.  

Disabled the SSE options for the iPod build as they are obviously Intel only.  Re-enabled these for the Mac build.

Disabled the auto vectorization optimisation on both IPod and Mac builds as this is for PowerPC architectures only.  

Disabled the model tuning option as this is PowerPC only.

Added the IPod deployment target option.

Enabled the position dependent code generation for the projects.  This should produce slightly faster code as it removes extra indirections being generated that are not necessary for an application. 

Re-enabled the optimisation level for the IPod such that debug uses none (0) and release uses smallest + fastest (s).

Finally fixed (I think!) the problem with the optimisation level causing the release build to not work on the IPod.  

Added an ENABLE_DEBUG_TRACE define to be able to switch on the debug break, assert and printing code regardless of whether we are using a release or debug build.  At the moment this is still specified in the TLDebug files but in the long term it would be nice to get this working as a runtime option with as little overhead of the debug code as possible.