Updated the AGON library to version 1.5.1 that is now built against the latest SDK (OS 3.2 SDK).
Removed the old AGON libraries - there is now only one AGON library that we link to by setting the appropriate base SDK and target deployment as well as weak linking to the MessageUI framework. This isn't actually as difficult as it first appeared. :)

Removed the -all_load build option as this caused issues when building some of our own libraries against the later SDK's and also linking with the AGON lib.  If we need to explicitly load symbols from a library use the -force_load option instead.

Removed some redundant files from the TootleSocial library.

Updated the social networking policy classes to use a TMessage to initialise them.  This is to bring it more in line with our other engine systems but also because of the varying amount of data needed to setup the different social networking libraries.  It also means we can potentially move this data into an XML file rather than hard coding it at some point. :)

Renamed the social networking BeginSession to CreateSession which now takes a TMessage as part of it's parameters and added a DestroySession - these handle construction and destruction of the social networking interfaces as required. 

Also added new BeginSession and EndSession routines.  This is mainly down to AGON using a two state system where you create the interface and then use a startgamesession/endgamesesssion pair during gameplay to control when scores are cached and synced.  Seeing as we use AGON in production I've made our classes match this for now.

Added skeleton files for support of the 'Game Center' framework that will be coming in iPhone OS 4.0.