Enabled the risky copy constructor for the TPtrArray.  This potentially calls a virtual functions so *could* be bad but alas it now fixes two issues, one for the widgets not being dereferenced correctly and one for the save options where when I duplicate the options asset TBinaryTree data using CopyTreeData it would corrupt the new copy when the asset is removed due to the TPtrArray not being setup correctly.
Re-instated the PC quit functionality when a window is closed.  Feel free to move/change this if it's not in an appropriate place so long as you can still close the window and the app shutdown.

Temporarily removed the group removal for the widget manager. Even with the TPtrArray copy constructor in which fixes one problem, there is still an issue where when the TKeyArray is re-ordered (sorted) it will not deallocate the TPtrArrays properly.

Changed the TApplication GetOptionsRef to use the applications name by default rather than the "bootup" ref.  I found I was simply specialising this on the applications using the same name as the app for the options file.

Added preliminary work on a TOptions class that is simply a wrapper to a TBinaryTree class for the options on the app.  It will eventually provide some extra routines for setting options and flagging the data as changed so it can be saved.

Setup the creation of an options asset if one doesn't exist when exiting a level as well as now copying the options data back into the asset when being saved.