Tidied up a few uses of TArray<TYPE> where they should be THeapArray or TPointerArray instead.  This fixes a few of the PC build errors for the Seaman Count and Rocket projects.
Added a FindNode routine to the TGraph class that will take a template parameter to convert the return node pointer to the that type using a reinterpret cast.

Updated the TPtrArray copy constructor and =operator to be more explicit about the type parameter(s).  This fixes an issue where the copy wasn't actually being called and consequently causing some crashes due to objects being deleted prematurely.